Controller.Chromecast/requirements.txt
flask_cors
flask-pymongo
fastapi
requests
gevent
python-kasa
pychromecast
python-icc